requirements.txt
astroid==1.6.4
bashlex==0.12
enum34==1.1.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.1
PyYAML==3.12
ruamel.yaml==0.15.37
six==1.11.0
wrapt==1.10.11
